+++ src/main/java/com/zea7ot/leetcode/lvl4/lc0051/SolutionApproach0Backtrack1.java
@@ -6,6 +6,7 @@
  * 
  * References:
  *  https://leetcode.com/problems/n-queens/discuss/19808/Accepted-4ms-c%2B%2B-solution-use-backtracking-and-bitmask-easy-understand.
+ *  https://mp.weixin.qq.com/s?__biz=MzAxODQxMDM0Mw==&mid=2247484709&idx=1&sn=1c24a5c41a5a255000532e83f38f2ce4&chksm=9bd7fb2daca0723be888b30345e2c5e64649fc31a00b05c27a0843f349e2dd9363338d0dac61&scene=178&cur_album_id=1318883740306948097#rd
  */
 package com.zea7ot.leetcode.lvl4.lc0051;
