+++ src/main/java/com/zea7ot/leetcode/lvl5/lc0037/SolutionApproach0Backtrack.java
@@ -7,6 +7,7 @@
  * 
  * References:
  *  https://leetcode.com/problems/sudoku-solver/discuss/15911/Less-than-30-line-clean-java-solution-using-DFS
+ *  https://mp.weixin.qq.com/s?__biz=MzAxODQxMDM0Mw==&mid=2247485097&idx=1&sn=a5e82da8646cd8985de6b2b0950de4e2&chksm=9bd7f8a1aca071b7b72e23013bc2a7c528ee913fded9278e2058bc98d7c746e439737d7abb5b&scene=178&cur_album_id=1318883740306948097#rd
  */
 package com.zea7ot.leetcode.lvl5.lc0037;
