+++ src/main/java/com/zea7ot/leetcode/tags/by_algorithm/dynamic_programming/resource/resources.txt
@@ -19,4 +19,5 @@
 ## Buying and selling stocks
 
     https://labuladong.gitbook.io/algo/di-ling-zhang-bi-du-xi-lie/tuan-mie-gu-piao-wen-ti
+    https://mp.weixin.qq.com/mp/appmsgalbum?__biz=MzAxODQxMDM0Mw==&action=getalbum&album_id=1318881141113536512&scene=173&from_msgid=2247484731&from_itemidx=1&count=3#wechat_redirect
