+++ src/main/java/com/zea7ot/leetcode/tags/by_algorithm/search/binary_search/summary.txt
@@ -94,6 +94,7 @@ This serves the list of all the LC problems related to Binary Search, regardless
 Resources:
     https://www.topcoder.com/community/competitive-programming/tutorials/binary-search/
     https://leetcode.com/problems/k-th-smallest-prime-fraction/discuss/115819/Summary-of-solutions-for-problems-%22reducible%22-to-LeetCode-378
+    https://mp.weixin.qq.com/s?__biz=MzAxODQxMDM0Mw==&mid=2247485044&idx=1&sn=e6b95782141c17abe206bfe2323a4226&chksm=9bd7f87caca0716aa5add0ddddce0bfe06f1f878aafb35113644ebf0cf0bfe51659da1c1b733&scene=21#wechat_redirect
 
     # Termination Conditions: while(left < right) vs while(left <= right)
     https://stackoverflow.com/a/30928332/6061609
